refactor: Extract roll test infrastructure helpers and table-driven tests
Promotes VALID_OUTCOMES and RollOutcome type to foundry/src/types/roll.ts
for reuse across all roll test files. Extracts getFirstMockCallArg and
setupGlobalMocks helpers to foundry/src/__mocks__/test-utils.ts.
Centralizes ChatMessage mock scaffolding in foundry/src/__mocks__/chat-message-mocks.ts.
Refactors skill-roll-integration.test.ts to use table-driven test pattern,
reducing duplication and improving maintainability across skill/franchise combinations.

Fixes #478, #468, #469, #470, #471
