+++ foundry/src/types/roll.ts
@@ -1,0 +1,3 @@
+export type RollOutcome = "good" | "partial" | "bad";
+
+export const VALID_OUTCOMES: RollOutcome[] = ["good", "partial", "bad"];
